add test cover

+++ messagequeue/message/message_test.go
@@ -9,8 +9,8 @@ import (
 func TestMessage(t *testing.T) {
 	m := NewMessage("hello", 10)
 	assert.Equal(t, "hello", m.Content)
-	assert.Equal(t, 10, m.ExpirySeconds)
-	assert.Equal(t, 0, m.LivedSeconds)
+	assert.Equal(t, int32(10), m.ExpirySeconds)
+	assert.Equal(t, int32(0), m.LivedSeconds)
 	assert.NotEqual(t, "", m.Id)
 }
